+++ feed/tables.py
@@ -1,6 +1,8 @@
 from sqlalchemy import MetaData, Integer, Column, Date, Enum, String, text, Index, Table
+from flask_sqlalchemy import SQLAlchemy
 
 metadata = MetaData()
+db= SQLAlchemy()
 
 arXiv_updates = Table(
     'arXiv_updates', metadata,
@@ -12,3 +14,17 @@ arXiv_updates = Table(
     Column('category', String(20), index=True),
     Index('document_id', 'document_id', 'date', 'action', 'category', unique=True)
 )
+
+class ArXivUpdate(db.Model):
+    __table__ = arXiv_updates
+    # Define the composite primary key
+    __table_args__ = (
+        db.PrimaryKeyConstraint('document_id', 'date', 'action', 'category'),
+    )
+    # Explicitly specify primary key columns
+    __mapper_args__ = {
+        'primary_key': [arXiv_updates.c.document_id, arXiv_updates.c.date, arXiv_updates.c.action, arXiv_updates.c.category]
+    }
+    # You can define indexes and additional attributes here
+    def __repr__(self):
+        return f"ArXivUpdate(document_id={self.document_id}, version={self.version}, action={self.action}, date={self.date}, category={self.category})"
